Split Rust and CPP implementation and make them optional
