+++ go/state/externalstate/external_state_test.go
@@ -8,6 +8,8 @@
 // On the date above, in accordance with the Business Source License, use of
 // this software will be governed by the GNU Lesser General Public License v3.
 
+//go:build carmen_cpp || carmen_rust
+
 package externalstate
 
 import (
